+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v0.5.1
 
 - refactoring (interface manager)
-- adding formatter widgets
+- adding formatter widgets, merge formatter options with column definition formatter options
+- adding advancved formatter (formatter widget per column)